+++ webservice/src/main/java/com/hortonworks/iotas/webservice/IotasApplication.java
@@ -26,7 +26,6 @@ import com.hortonworks.iotas.storage.cache.writer.StorageWriteThrough;
 import com.hortonworks.iotas.storage.cache.writer.StorageWriter;
 import com.hortonworks.iotas.common.FileEventHandler;
 import com.hortonworks.iotas.common.exception.ConfigException;
-
 import com.hortonworks.iotas.streams.metrics.TimeSeriesQuerier;
 import com.hortonworks.iotas.service.CatalogService;
 import com.hortonworks.iotas.service.FileWatcher;
@@ -34,8 +33,6 @@ import com.hortonworks.iotas.storage.CacheBackedStorageManager;
 import com.hortonworks.iotas.storage.Storable;
 import com.hortonworks.iotas.storage.StorableKey;
 import com.hortonworks.iotas.storage.StorageManager;
-import com.hortonworks.iotas.storage.impl.jdbc.JdbcStorageManager;
-import com.hortonworks.iotas.storage.impl.memory.InMemoryStorageManager;
 import com.hortonworks.iotas.processor.CustomProcessorUploadHandler;
 import com.hortonworks.iotas.streams.catalog.service.StreamCatalogService;
 import com.hortonworks.iotas.streams.layout.storm.StormTopologyLayoutConstants;
@@ -113,14 +110,25 @@ public class IotasApplication extends Application<IotasConfiguration> {
 
     private StorageManager getCacheBackedDao(IotasConfiguration iotasConfiguration) {
         StorageProviderConfiguration storageProviderConfiguration = iotasConfiguration.getStorageProviderConfiguration();
-        final String providerType = storageProviderConfiguration.getType();
-        final StorageManager dao = providerType.equalsIgnoreCase(JDBC) ?
-                JdbcStorageManager.createStorageManager(storageProviderConfiguration.getProperties()) : new InMemoryStorageManager();
+        final StorageManager dao = getStorageManager(storageProviderConfiguration);
         final CacheBuilder cacheBuilder = getGuavaCacheBuilder();
         final Cache<StorableKey, Storable> cache = getCache(dao, cacheBuilder);
         final StorageWriter storageWriter = getStorageWriter(dao);
 
         return doGetCacheBackedDao(cache, storageWriter);
+    }
+
+    private StorageManager getStorageManager(StorageProviderConfiguration storageProviderConfiguration) {
+        final String providerClass = storageProviderConfiguration.getProviderClass();
+        StorageManager storageManager = null;
+        try {
+            storageManager = (StorageManager) Class.forName(providerClass).newInstance();
+        } catch (InstantiationException | IllegalAccessException | ClassNotFoundException e) {
+            throw new RuntimeException(e);
+        }
+        storageManager.init(storageProviderConfiguration.getProperties());
+
+        return storageManager;
     }
 
     private StorageWriter getStorageWriter(StorageManager dao) {